SDC move sdcavr.* into sdc.cpp  No functional changes
